Moves root reducer into reducers/ directory
